+++ lib/ui/core_ui.dart
@@ -15,16 +15,6 @@ class CoreUi {
     print("godot-cli version $packageVersion");
   }
 
-  void printAccessDenied() {
-    String roleHint = "(as Administrator)";
-
-    if (Platform.isLinux || Platform.isMacOS) {
-      roleHint = "(as root)";
-    }
-    stderr.writeln("${"[X]".red} Access denied");
-    stderr.writeln("    ${"·".red} ${"You must run this program using elevated privileges $roleHint.".bold}");
-  }
-
   void printFirstRun() {
     print("${"[!]".yellow} Command denied");
     print("    ${"·".yellow} ${"Please run 'doctor' first.".bold}");